% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{aut}
\alias{aut}
\title{AUT Summary Statistics}
\format{A data frame with 11 variables:
\describe{
\item{\code{snpid}}{SNP rs identifier}
\item{\code{hg18chr}}{hg18 chromosome (1-22)}
\item{\code{bp}}{hg18 base position of SNP}
\item{\code{a1}}{reference allele (not necessarily minor allele)}
\item{\code{a2}}{alternate allele}
\item{\code{or}}{odds ratio from the logistic regression including PCA
covariates}
\item{\code{se}}{standard error of the odds ratio}
\item{\code{pval}}{asymptotic p-value}
\item{\code{info}}{INFO score from imputation, ratio of variances can exceed
1}
\item{\code{ngt}}{number of studies in which this SNP is directly genotyped
(not imputed)}
\item{\code{CEUaf}}{frequency of `a1` in HapMap3 CEU (HapMap2 for BIP)}
}}
\source{
\url{https://www.med.unc.edu/pgc/results-and-downloads}
}
\usage{
aut
}
\description{
Summary statistics for the autism spectrum disorder (AUT) subset of PGC's
cross disorder analysis.
}
\references{
Cross-disorder group of the Psychiatric Genomics Consortium, Identification of risk loci with shared effects on five major psychiatric disorders: a genome-wide analysis. Lancet, epub (2013). DOI: \href{https://doi.org/10.1016/S0140-6736(12)62129-1}{10.1016/S0140-6736(12)62129-1}.
}
\seealso{
Other PGC datasets: \code{\link{adhd}}, \code{\link{bip}},
  \code{\link{mdd}}, \code{\link{scz}}
}
\keyword{datasets}
