% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bip}
\alias{bip}
\title{BIP Summary Statistics}
\format{A data frame with 11 variables:
\describe{
\item{\code{snpid}}{SNP rs identifier}
\item{\code{hg18chr}}{hg18 chromosome (1-22)}
\item{\code{bp}}{hg18 base position of SNP}
\item{\code{a1}}{reference allele (not necessarily minor allele)}
\item{\code{a2}}{alternate allele}
\item{\code{or}}{odds ratio from the logistic regression including PCA
covariates}
\item{\code{se}}{standard error of the odds ratio}
\item{\code{pval}}{asymptotic p-value}
\item{\code{info}}{INFO score from imputation, ratio of variances can exceed
1}
\item{\code{ngt}}{number of studies in which this SNP is directly genotyped
(not imputed)}
\item{\code{CEUaf}}{frequency of `a1` in HapMap3 CEU (HapMap2 for BIP)}
}}
\source{
\url{https://www.med.unc.edu/pgc/results-and-downloads}
}
\usage{
bip
}
\description{
Summary statistics for the bipolar disorder (BIP) subset of PGC's
cross disorder analysis.
}
\references{
Psychiatric GWAS Consortium Bipolar Disorder Working Group, Large-scale genome-wide association analysis of bipolar disorder identifies a new susceptibility locus near ODZ4. Nature Genetics 43, 977 (2011). DOI: \href{https://doi.org/10.1038/ng.943}{10.1038/ng.943}.
}
\seealso{
Other PGC datasets: \code{\link{adhd}}, \code{\link{aut}},
  \code{\link{mdd}}, \code{\link{scz}}
}
\keyword{datasets}
