% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mdd}
\alias{mdd}
\title{mdd Summary Statistics}
\format{A data frame with 11 variables:
\describe{
\item{\code{snpid}}{SNP rs identifier}
\item{\code{hg18chr}}{hg18 chromosome (1-22)}
\item{\code{bp}}{hg18 base position of SNP}
\item{\code{a1}}{reference allele (not necessarily minor allele)}
\item{\code{a2}}{alternate allele}
\item{\code{or}}{odds ratio from the logistic regression including PCA
covariates}
\item{\code{se}}{standard error of the odds ratio}
\item{\code{pval}}{asymptotic p-value}
\item{\code{info}}{INFO score from imputation, ratio of variances can exceed
1}
\item{\code{ngt}}{number of studies in which this SNP is directly genotyped
(not imputed)}
\item{\code{CEUaf}}{frequency of `a1` in HapMap3 CEU (HapMap2 for BIP)}
}}
\source{
\url{https://www.med.unc.edu/pgc/results-and-downloads}
}
\usage{
mdd
}
\description{
Summary statistics for the major depressive disorder (mdd) subset of PGC's
cross disorder analysis.
}
\references{
Major Depressive Disorder Working Group of the PGC, A mega-analysis of genome-wide association studies for major depressive disorder. Molecular Psychiatry. (2013). DOI: \href{https://doi.org/10.1038/mp.2012.21}{10.1038/mp.2012.21}.
}
\seealso{
Other PGC datasets: \code{\link{adhd}}, \code{\link{aut}},
  \code{\link{bip}}, \code{\link{scz}}
}
\keyword{datasets}
