% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pgc}
\alias{pgc}
\alias{adhd}
\alias{aut}
\alias{bip}
\alias{mdd}
\alias{scz}
\title{PGC Cross Disorder Summary Statistics}
\format{A data frame with 11 variables:
\itemize{
\item \strong{\code{snpid}}: SNP rs identifier
\item \strong{\code{hg18chr}}: hg18 chromosome (1-22)
\item \strong{\code{bp}}: hg18 base position of SNP
\item \strong{\code{a1}}: reference allele (not necessarily minor allele)
\item \strong{\code{a2}}: alternate allele
\item \strong{\code{or}}: odds ratio from the logistic regression including PCA covariates
\item \strong{\code{se}}: standard error of the odds ratio
\item \strong{\code{pval}}: asymptotic p-value
\item \strong{\code{info}}: INFO score from imputation, ratio of variances can exceed 1
\item \strong{\code{ngt}}: number of studies in which this SNP directly genotyped (not imputed)
\item \strong{\code{CEUaf}}: frequency of \code{a1} in HapMap3 CEU (HapMap2 for BIP)
}}
\source{
https://www.med.unc.edu/pgc/results-and-downloads
}
\usage{
adhd

aut

bip

mdd

scz
}
\description{
Summary statistics for the 5 individual disorders used in the Psychiatric
Genomics Consortium (PGC) cross disorder analysis, including autism spectrum
disorder (\code{aut}), attention deficit-hyperactivity disorder (\code{adhd}), bipolar
disorder (\code{bip}), major depressive disorder (\code{mdd}), and schizophrenia
(\code{scz}).
}
\references{
\enumerate{
\item Neale et al., Meta-analysis of genome-wide association studies of
attention- deficit/hyperactivity disorder. J. Am. Acad. Child Adolesc.
Psychiatry 49, 884 (Sep, 2010).
\item Psychiatric GWAS Consortium Bipolar Disorder Working Group, Large-scale
genome-wide association analysis of bipolar disorder identifies a new
susceptibility locus near ODZ4. Nature Genetics 43, 977 (2011).
\item Major Depressive Disorder Working Group of the PGC, A mega-analysis of
genome-wide association studies for major depressive disorder. Molecular
Psychiatry. (2013). DOI: 10.1038/mp.2012.21.
\item Schizophrenia PGC, Genome-wide association study of schizophrenia
identifies five novel loci. Nature Genetics 43, 969 (2011).
\item Cross-disorder group of the Psychiatric Genomics Consortium,
Identification of risk loci with shared effects on five major psychiatric
disorders: a genome-wide analysis. Lancet, epub (2013).
}
}
\keyword{datasets}
