% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgcxd.r
\docType{package}
\name{pgcxd}
\alias{pgcxd}
\alias{pgcxd-package}
\title{PGC Cross Disorder Summary Statistics}
\source{
https://www.med.unc.edu/pgc
}
\description{
Summary statistics for the 5 individual disorders used in the Psychiatric
Genomics Consortium (PGC) cross disorder analysis, including autism spectrum
disorder (\code{aut}), attention deficit-hyperactivity disorder
(\code{adhd}), bipolar disorder (\code{bip}), major depressive disorder
(\code{mdd}), and schizophrenia (\code{scz}).
}
\references{
Cross-disorder group of the Psychiatric Genomics Consortium,
Identification of risk loci with shared effects on five major psychiatric
disorders: a genome-wide analysis. Lancet, epub (2013).
}
